Remove the module installation hack.
Install by default in the Linux-PAM path (/lib/security), will add an
override to configure.ac later, and in the mean time OpenPAM (FreeBSD) can
do it in different way.

Using the uname or host_os value is *not* the right way to choose where to
install, for instance this will be wrong for multilib-capable distributions
such as Gentoo Linux as it is.