Improve the check for libpam_misc.
Instead of assuming that Linux-PAM is only used on Linux, and OpenPAM on
FreeBSD (both of which are not the case: the former is used on OSX and
FreeBSD 4.x, the latter is used on some Linux distributions), make sure to
find the proper conversation, between the Linux-PAM pam_misc-based one or
the default-provided OpenPAM one.

Also, the module does not require the conversation function to begin with,
so if it's not found, simply ignore the two “tests” instead of failing or
something like that.