query_acct should return two or three columns
As presented by James Bellinger in #2940609 the check for number of
fields returned from query_acct, in this case >= 4, is useless,
therefore as the way it is implemented the pam expects two or three
columns, the first if account is expired, second one if a new password
should be asked and the third (optional) if the password is NULL or
blank '' (if so a new password required).
Code has been refactored to become a little bit more clear but it should
be checked in the future.