Use the Autoconf Archive-provided macro to look for PostgreSQL.
Instead of just checking for libpq and for the libpq-fe.h header files, use
the macro that checks for and abides to pg_config. This way it's possible
to auto-detect installation paths on all distributions (still allowing them
to change the pg_config script) even when they don't install in “default”
paths.