Stop hardcoding the path to the configuration file, make it configurable.
Instead of just always using /etc/pam_pgsql.conf as configuration file,
allow configuration of the directory where to find the configuration
file. Let the default be /etc for any prefix given, with the user able to
override the location.

This is particularly important for distributions using Linux-PAM, as most
of the PAM modules are configured in /etc/security rather than simply /etc.