Implement visibility support for the PAM module.
If the compiler supports the -fvisibility flag and the default visibility
attribute, make use of the ELF visibility support so that only the PAM
interface is exported by the module and none of the other symbols.

This also ensures that the (generic) symbols for MD5 and SHA1 are not going
to conflict with an eventual implementation loaded by the original process.