Use libgcrypt for MD5 and SHA1 hashes.
Instead of using the included code for MD5 and SHA1, which turned out to
not be entirely tested on non-x86 architectures, rely on a tested and
supported library to do the work.

The libgcrypt library makes it very easy and fast to use the hashing
functions, and is widely used by other projects as well.